requirements.txt
asgiref==3.8.1
astroid==3.3.8
certifi==2024.12.14
charset-normalizer==3.4.1
dill==0.3.9
Django==4.2.17
django-cors-headers==4.6.0
django-debug-toolbar==5.0.1
django-extensions==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
hubspot==0.1.14.dev0
hubspot-api==0.7.0
hubspot-api-client==11.1.0
idna==3.10
isort==6.0.0
mccabe==0.7.0
mysql-connector-python==9.1.0
pillow==11.0.0
platformdirs==4.3.6
PyJWT==2.10.1
pylint==3.3.4
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
simplejson==3.19.3
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==1.26.16
